Implement Nova traits for Pallas.
