+++ Cargo.toml
@@ -7,8 +7,14 @@ edition = "2018"
 [dependencies]
 crossbeam = "0.8"
 ff = "0.10"
-pasta_curves = "0.1.2"
+merlin = "2.0.0"
+#nova = { package="nova-snark", version = "0.1", git = "https://github.com/microsoft/nova"}
+nova = { package="nova-snark", version = "0.1", git = "https://github.com/porcuquine/nova", branch="upgrade-rand"}
+#pasta_curves = "0.1.2"
+pasta_curves = { git = "https://github.com/porcuquine/pasta_curves", branch = "reexport-group" }
+#rand = "0.7.3"
 rand = "0.8"
+subtle = "2.3"
 
 [dev-dependencies]
 criterion = "0.3"
