+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.54.0
+nightly-2021-08-31
